refactor: move `clear_pending_proposal` from `CentralContext` to `ConversationGuard`
